refactor(sync): move utility functions to separate modules
- Move ChangelogFile interface definition to remove ResolvedVersion extension
- Extract demo parsing logic to scripts/extractors/demos.ts module
- Move logging utilities to scripts/utils/log.ts module
- Extract markdown parsing functions to scripts/extractors/markdown.ts module
- Move semantic structure reading to scripts/extractors/semantic-structure.ts module
- Extract token processing functions to scripts/extractors/tokens.ts module
- Remove duplicate type definitions from sync.ts script
- Consolidate related functionality into dedicated extractor modules
